backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
twilio==7.16.0
geopy==2.2.0
python-dotenv==0.19.0
scikit-learn==0.24.2
numpy==1.21.0
scipy==1.7.0
joblib==1.0.1
tensorflow==2.13.0
pandas==1.5.3